backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg2-binary==2.9.9
pandas==2.1.3
openpyxl==3.1.2
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0